requirements.txt
fastapi
python-multipart
pydantic
pypdf

# llm
# langchain-ollama
# langchain-groq
google-generativeai==0.3.2

langchain-community
langchain-text-splitters
langchain-chroma
langchain-core
langchain
langchain-huggingface
python-dotenv
chromadb
huggingface-hub

# sqlalchemy stuff
sqlalchemy 
psycopg2-binary 
passlib[bcrypt]==1.7.4

# auth stuff
python-jose[cryptography] 
python-multipart
bcrypt==4.0.1


